+++ lib/password_strength_checker.dart
@@ -1,4 +1,13 @@
 /// [PasswordStrengthChecker] is a [Widget] that checks the strength of a password.
 library password_strength_checker;
 
+// UI
 export 'src/password_strength_checker.dart';
+export 'src/password_strength_form_checker.dart';
+// Configurations
+export 'src/config/password_strength_checker_configuration.dart';
+export 'src/config/text_form_field_configuration.dart';
+// Models
+export 'src/dict/common_passwords.dart';
+// Utils
+export 'src/enum/password_strength_enum.dart';