+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.1.0
+- Implemented **PasswordStrengthFormChecker** and moved all **PasswordStrengthChecker** properties into **PasswordStrengthCheckerConfiguration** object.
+- Implemented a *commonDictionary* to get the first 10,000 common passwords.
+- Extended **PasswordStrength** with *alreadyExposed*.
 ## 1.0.1
 Updated docs.
 ## 1.0.0